refactor(invoice-total-calculator): implement SRP, monetary representation to int

+++ 01-report-processing/03-invoice-total-calculator/src/Contracts/SubtotalCalculator.php
@@ -7,5 +7,5 @@ use App\Models\Invoice;
 
 interface SubtotalCalculator
 {
-	public function calculate(Invoice $invoice): float;
+	public function calculate(Invoice $invoice): int;
 }
